Move ephemeral users session check to a concern and improve flow
